Switch to a monorepo setup with lerna
Closes #23
