+++ package.json
@@ -1,47 +1,7 @@
 {
-  "name": "tegel",
-  "version": "0.4.0",
-  "description": "A simple feature toggle service for Node.js",
-  "main": "lib/index.js",
-  "scripts": {
-    "test": "jest"
-  },
-  "keywords": [
-    "feature",
-    "flag",
-    "toggle",
-    "switch",
-    "express",
-    "feature flag",
-    "feature toggle",
-    "feature switch"
-  ],
-  "author": {
-    "name": "Fatih Altinok",
-    "email": "fatihaltinok@live.com",
-    "url": "https://6nok.org"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git@github.com:frontsideair/tegel"
-  },
-  "homepage": "http://github.com/frontsideair/tegel",
-  "bugs": "http://github.com/frontsideair/tegel/issues",
-  "files": [
-    "lib"
-  ],
-  "license": "MIT",
-  "dependencies": {
-    "@keyv/sqlite": "^2.0.1",
-    "body-parser": "^1.19.0",
-    "express": "^4.17.1",
-    "keyv": "^4.0.0"
-  },
+  "name": "root",
+  "private": true,
   "devDependencies": {
-    "jest": "^24.9.0",
-    "supertest": "^4.0.2"
-  },
-  "engines": {
-    "node": ">=8.0.0"
+    "lerna": "^3.20.2"
   }
 }
