+++ packages/tegel/index.js
@@ -1,7 +1,3 @@
-var express = require("express");
-var bodyParser = require("body-parser");
-var view = require("./view");
-
 class Tegel {
   constructor({ toggles, storage, interval = 1000 }) {
     this.toggles = toggles;
@@ -42,28 +38,4 @@ function middleware(tegel) {
   };
 }
 
-function router(tegel) {
-  var app = express();
-  app.use(bodyParser.urlencoded({ extended: false }));
-
-  function toggleData() {
-    return tegel.toggles.map(toggle => [toggle, tegel.get(toggle.name)]);
-  }
-
-  app.get("/", async (req, res) => {
-    res.send(view(toggleData()));
-  });
-
-  app.post("/", async (req, res) => {
-    await Promise.all(
-      tegel.toggles.map(toggle =>
-        tegel.set(toggle.name, req.body[toggle.name] === "on" ? true : false)
-      )
-    );
-    res.send(view(toggleData()));
-  });
-
-  return app;
-}
-
-module.exports = { Tegel, middleware, router };
+module.exports = { Tegel, middleware };
